# ROB: Fix underflow bug and reduce allocations 